+++ src/Designer/frontend/settings/features/orgs/routes/RoutePaths.ts
@@ -1,5 +1,4 @@
 export enum RoutePaths {
-  Org = 'orgs/:org',
-  ContactPoints = 'contact-points',
   BotAccounts = 'bot-accounts',
+  ContactPoints = 'contact-points',
 }