+++ src/Runtime/localtest/src/Configuration/LocalPlatformSettings.cs
@@ -14,11 +14,6 @@ namespace LocalTest.Configuration
         /// </summary>
         public string LocalTestingStorageBasePath { get; set; }
 
-        /// <summary>
-        /// The folder where the app that is tested is located. Used to retrieve app configuration
-        /// </summary>
-        public string AppRepositoryBasePath { get; set; }
-
         public string BlobStorageFolder { get; set; } = "blobs/";
 
         public string NotificationsStorageFolder { get; set; } = "notifications/";
@@ -42,16 +37,6 @@ namespace LocalTest.Configuration
                 _localTestDataPath = value;
             }
         }
-
-        /// <summary>
-        /// Url for the local app when LocalAppMode == http
-        /// <summary>
-        public string LocalAppUrl { get; set; }
-
-        /// <summary>
-        /// which access mode to use ("file", "http")
-        /// <summary>
-        public string LocalAppMode { get; set; }
 
         public string DocumentDbFolder { get; set; } = "documentdb/";
 
@@ -80,13 +65,5 @@ namespace LocalTest.Configuration
         public string TenorDataFolder { get; set; } = "tenorUsers";
 
         public string InstanceDelegationsDataFolder { get; set; } = "instanceDelegations";
-
-        public string LocalGrafanaUrl { get; set; }
-
-        public string LocalPdfServiceUrl { get; set; }
-
-        public string LocalWorkflowEngineUrl { get; set; }
-
-        public string LocalPgAdminUrl { get; set; }
     }
 }